Start revisions of UI Tool layout
