+++ developer-preview/src/app/SelectValues.tsx
@@ -1,0 +1,12 @@
+import { UI } from "~/app/UI";
+
+export const SelectValues = () => {
+  return (
+    <div>
+      <UI.HeadingField>Preview with</UI.HeadingField>
+      <UI.Select disabled onChange={() => null}>
+        <UI.Option value="">Placeholder values</UI.Option>
+      </UI.Select>
+    </div>
+  );
+};
